backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-contrib-python-headless==4.8.1.78
numpy==1.26.4
aiofiles==23.2.1
pydantic==2.4.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.0 